refactor: complete loadAssignmentAndSetup adoption and remaining file splits
The last items from docs/audit-2026-06.md:
- All 15 remaining inline assignment-lookup chains in the web layer now
  go through loadAssignment / loadAssignmentAndSetup (dashboard cluster,
  BrightSpace, course-section move). Only the MCP tools still resolve
  by ID from tool arguments, which is correct for that surface.
- StudentCourseRoutes+History: one resolveStudentAssignmentAction
  preamble + redirectToStudentSubmissions helper replace seven copies
  (one defense-in-depth instructor check kept in the shared helper).
- Splits: WebRoutes+Submission 989->479 (SubmissionResultPresenter),
  CourseBundleRoutes 801->346 (+Import), InstructorDashboardRoutes+
  Submissions 850->218 (+GradesCSV, +StudentActions);
  preferredResultsBySubmissionID moved to Helpers/.

Verified: build, 103 tests in 8 suites, SwiftLint --strict 0 violations
in 599 files, swift-format clean.
