+++ Sources/APIServer/Routes/Web/CourseBundleRoutes.swift
@@ -12,13 +12,8 @@
 //   testsetups/<id>.zip      — instructor test-setup archives
 //   submissions/<id>.<ext>   — student submission files
 //
-// Import rules:
-//   - Same course code, active     → reject with error message
-//   - Same course code, archived   → create a second course (admin can rename)
-//   - Unknown course code          → create fresh
-//   - Users: match by username or create placeholder (inert until password reset)
-//   - All DB IDs are regenerated; bundleIDs are internal cross-references only.
-//   - validationStatus is NOT imported; assignments land as "pending" validation.
+// This file holds boot + the export pipeline; the import handler and its
+// phase helpers live in CourseBundleRoutes+Import.swift.
 
 import Core
 import Fluent
@@ -328,231 +323,6 @@ struct CourseBundleRoutes: RouteCollection {
 
         return Response(status: .ok, headers: headers, body: .init(data: zipData))
     }
-
-    // MARK: - POST /admin/courses/import
-
-    @Sendable
-    func importCourse(req: Request) async throws -> View {
-        let caller = try req.auth.require(APIUser.self)
-        guard caller.isAdmin else { throw Abort(.forbidden) }
-
-        let fileBytes = try readUploadedBundleBytes(req: req)
-
-        let tmpZipPath = FileManager.default.temporaryDirectory
-            .appendingPathComponent("chickadee-import-\(UUID().uuidString).zip").path
-        let extractDir = FileManager.default.temporaryDirectory
-            .appendingPathComponent("chickadee-import-ex-\(UUID().uuidString)", isDirectory: true)
-
-        defer {
-            try? FileManager.default.removeItem(atPath: tmpZipPath)
-            try? FileManager.default.removeItem(at: extractDir)
-        }
-
-        try await extractUploadedBundle(
-            fileBytes: fileBytes, tmpZipPath: tmpZipPath, extractDir: extractDir)
-
-        let manifest = try parseBundleManifest(extractDir: extractDir)
-
-        try validateBundleFiles(manifest: manifest, extractDir: extractDir)
-
-        let setupsDir = req.application.testSetupsDirectory
-        let subsDir = req.application.submissionsDirectory
-
-        let tally = try await performImportTransaction(
-            db: req.db,
-            manifest: manifest,
-            extractDir: extractDir,
-            setupsDir: setupsDir,
-            subsDir: subsDir
-        )
-
-        await AuditLogger.record(
-            action: .courseBundleImported,
-            targetType: .course,
-            targetID: tally.courseID.uuidString,
-            metadata: [
-                "course_code": tally.courseCode,
-                "test_setups": String(tally.testSetupsImported),
-                "assignments": String(tally.assignmentsImported),
-                "submissions": String(tally.submissionsImported),
-            ],
-            on: req
-        )
-        return try await renderImportResult(req: req, tally: tally)
-    }
-
-    // ── 1. Receive the uploaded bundle ────────────────────────────────
-
-    private func readUploadedBundleBytes(req: Request) throws -> [UInt8] {
-        struct BundleUpload: Content {
-            let file: File
-        }
-        let upload = try req.content.decode(BundleUpload.self)
-        var buffer = upload.file.data
-        guard buffer.readableBytes > 0,
-            let fileBytes = buffer.readBytes(length: buffer.readableBytes)
-        else {
-            throw AppError.badRequest(reason: "Empty bundle upload")
-        }
-        return fileBytes
-    }
-
-    // ── 2. Save to temp file and extract ─────────────────────────────
-
-    private func extractUploadedBundle(
-        fileBytes: [UInt8], tmpZipPath: String, extractDir: URL
-    ) async throws {
-        try Data(fileBytes).write(to: URL(fileURLWithPath: tmpZipPath))
-        try await extractZipArchive(zipPath: tmpZipPath, into: extractDir)
-    }
-
-    // ── 3. Parse bundle.json ──────────────────────────────────────────
-
-    private func parseBundleManifest(extractDir: URL) throws -> CourseBundleManifest {
-        let bundleJSONPath = extractDir.appendingPathComponent("bundle.json")
-        guard let manifestData = try? Data(contentsOf: bundleJSONPath) else {
-            throw AppError.badRequest(reason: "bundle.json not found in archive")
-        }
-
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .iso8601
-        let manifest: CourseBundleManifest
-        do {
-            manifest = try decoder.decode(CourseBundleManifest.self, from: manifestData)
-        } catch {
-            throw AppError.invalidParameter(name: "bundle.json", reason: "\(error)")
-        }
-
-        guard manifest.schemaVersion == 1 else {
-            throw Abort(
-                .badRequest,
-                reason: "Unsupported bundle schemaVersion \(manifest.schemaVersion); expected 1")
-        }
-
-        return manifest
-    }
-
-    // ── 4. Validate all referenced files exist ────────────────────────
-
-    private func validateBundleFiles(
-        manifest: CourseBundleManifest, extractDir: URL
-    ) throws {
-        for setup in manifest.testSetups {
-            let path = extractDir.appendingPathComponent(setup.zipFilename)
-            guard FileManager.default.fileExists(atPath: path.path) else {
-                throw Abort(
-                    .badRequest,
-                    reason: "Bundle is missing test setup file: \(setup.zipFilename)")
-            }
-        }
-        for sub in manifest.submissions {
-            let path = extractDir.appendingPathComponent(sub.submissionFilename)
-            guard FileManager.default.fileExists(atPath: path.path) else {
-                throw Abort(
-                    .badRequest,
-                    reason: "Bundle is missing submission file: \(sub.submissionFilename)")
-            }
-        }
-    }
-
-    // ── 6. Transactional import ───────────────────────────────────────
-    // The conflict check (formerly step 5) is now the first thing inside the
-    // transaction so there is no outstanding req.db cursor before the transaction
-    // begins. On SQLite this prevents "busy: cannot commit transaction — SQL
-    // statements in progress" errors caused by an open cursor from a pre-transaction
-    // query lingering when the COMMIT fires.
-    //
-    // Returns a tally from the closure to avoid captured-var mutation warnings
-    // (errors in Swift 6 strict mode).
-
-    private func performImportTransaction(
-        db: Database,
-        manifest: CourseBundleManifest,
-        extractDir: URL,
-        setupsDir: String,
-        subsDir: String
-    ) async throws -> ImportTally {
-        try await db.transaction { (db) -> ImportTally in
-            // 6a. Check for course code conflicts (moved inside transaction)
-            let existingCourse = try await APICourse.query(on: db)
-                .filter(\.$code == manifest.course.code)
-                .first()
-            if let existing = existingCourse, !existing.isArchived {
-                throw Abort(
-                    .conflict,
-                    reason: """
-                        A course with code "\(manifest.course.code)" already exists and is active. \
-                        Archive it first, then re-import.
-                        """)
-            }
-
-            var t = ImportTally(
-                courseID: UUID(),
-                courseCode: manifest.course.code,
-                courseName: manifest.course.name
-            )
-
-            // 6b. Create course
-            let importedMode = bundledCourseEnrollmentMode(manifest.course)
-            let newCourse = APICourse(
-                code: manifest.course.code, name: manifest.course.name,
-                enrollmentMode: importedMode)
-            try await newCourse.save(on: db)
-            guard let newCourseID = newCourse.id else {
-                throw AppError.internalFailure(reason: "Created course missing id after save")
-            }
-            t.courseID = newCourseID
-            t.courseCode = newCourse.code
-            t.courseName = newCourse.name
-
-            // 6c. Resolve users → userIDMap[bundleID] = live UUID
-            let userIDMap = try await importBundledUsers(manifest: manifest, db: db, tally: &t)
-
-            // 6d. Create enrollments for enrolled users
-            try await importBundledEnrollments(
-                manifest: manifest, userIDMap: userIDMap, courseID: t.courseID, db: db)
-
-            // 6e. Create test setups → setupIDMap[bundleID] = new live ID
-            let setupIDMap = try await importBundledTestSetups(
-                manifest: manifest, extractDir: extractDir, setupsDir: setupsDir,
-                courseID: t.courseID, db: db, tally: &t)
-
-            // 6f. Create assignments
-            try await importBundledAssignments(
-                manifest: manifest, setupIDMap: setupIDMap, courseID: t.courseID,
-                db: db, tally: &t)
-
-            // 6g. Create submissions → subIDMap[bundleID] = new live ID
-            let subIDMap = try await importBundledSubmissions(
-                manifest: manifest, extractDir: extractDir, subsDir: subsDir,
-                idMaps: ImportIDMaps(userIDMap: userIDMap, setupIDMap: setupIDMap),
-                db: db, tally: &t)
-
-            // 6h. Create results
-            try await importBundledResults(
-                manifest: manifest, subIDMap: subIDMap, db: db, tally: &t)
-
-            return t
-        }
-    }
-
-    // ── 8. Render result page ─────────────────────────────────────────
-
-    private func renderImportResult(req: Request, tally: ImportTally) async throws -> View {
-        let ctx = ImportResultContext(
-            currentUser: req.currentUserContext,
-            courseID: tally.courseID.uuidString,
-            courseCode: tally.courseCode,
-            courseName: tally.courseName,
-            testSetupsImported: tally.testSetupsImported,
-            assignmentsImported: tally.assignmentsImported,
-            usersCreated: tally.usersCreated,
-            usersMatched: tally.usersMatched,
-            submissionsImported: tally.submissionsImported,
-            resultsImported: tally.resultsImported
-        )
-        return try await req.view.render("admin-import-result", ctx)
-    }
 }
 
 // MARK: - Export data carriers
@@ -573,229 +343,4 @@ private struct ExportBundleIDs {
     let setupBundleIDByID: [String: String]
     let assignBundleIDByID: [UUID: String]
     let subBundleIDByID: [String: String]
-}
-
-/// Bundle-id → live-DB-id maps built up during the import transaction.
-private struct ImportIDMaps {
-    let userIDMap: [String: UUID]
-    let setupIDMap: [String: String]
-}
-
-// MARK: - Transaction tally
-
-/// Mutable counters accumulated inside the import transaction and returned to the caller.
-/// Using a local `var` inside the closure and returning it avoids the Swift 6
-/// "mutation of captured var in concurrently-executing code" error.
-private struct ImportTally: Sendable {
-    var courseID: UUID
-    var courseCode: String
-    var courseName: String
-    var usersCreated: Int = 0
-    var usersMatched: Int = 0
-    var testSetupsImported: Int = 0
-    var assignmentsImported: Int = 0
-    var submissionsImported: Int = 0
-    var resultsImported: Int = 0
-}
-
-// MARK: - View context
-
-private struct ImportResultContext: Encodable {
-    let currentUser: CurrentUserContext?
-    let courseID: String
-    let courseCode: String
-    let courseName: String
-    let testSetupsImported: Int
-    let assignmentsImported: Int
-    let usersCreated: Int
-    let usersMatched: Int
-    let submissionsImported: Int
-    let resultsImported: Int
-}
-
-// MARK: - Import phase helpers (6c–6h)
-//
-// These are fileprivate free functions rather than methods on `CourseBundleRoutes`
-// so the route struct stays under the swiftlint type_body_length limit.
-
-private func importBundledUsers(
-    manifest: CourseBundleManifest, db: Database, tally: inout ImportTally
-) async throws -> [String: UUID] {
-    var userIDMap: [String: UUID] = [:]
-    for bundledUser in manifest.users {
-        if let existing = try await APIUser.query(on: db)
-            .filter(\.$username == bundledUser.username)
-            .first()
-        {
-            guard let existingID = existing.id else {
-                throw AppError.internalFailure(reason: "User '\(bundledUser.username)' missing id")
-            }
-            userIDMap[bundledUser.bundleID] = existingID
-            tally.usersMatched += 1
-        } else {
-            // Create placeholder — inert until password reset or SSO login.
-            let newUser = APIUser(
-                username: bundledUser.username,
-                passwordHash: "",  // inert placeholder
-                role: bundledUser.role,
-                authProvider: nil,
-                email: bundledUser.email,
-                displayName: bundledUser.displayName
-            )
-            try await newUser.save(on: db)
-            guard let newUserID = newUser.id else {
-                throw AppError.internalFailure(reason: "Created user missing id after save")
-            }
-            userIDMap[bundledUser.bundleID] = newUserID
-            tally.usersCreated += 1
-        }
-    }
-    return userIDMap
-}
-
-private func importBundledEnrollments(
-    manifest: CourseBundleManifest,
-    userIDMap: [String: UUID],
-    courseID: UUID,
-    db: Database
-) async throws {
-    for bundleID in manifest.enrolledUserBundleIDs {
-        guard let uid = userIDMap[bundleID] else { continue }
-        // Skip if already enrolled (matched user already in another course).
-        let alreadyEnrolled = try await APICourseEnrollment.query(on: db)
-            .filter(\.$userID == uid)
-            .filter(\.$course.$id == courseID)
-            .first()
-        if alreadyEnrolled == nil {
-            let enrollment = APICourseEnrollment(userID: uid, courseID: courseID)
-            try await enrollment.save(on: db)
-        }
-    }
-}
-
-private func importBundledTestSetups(
-    manifest: CourseBundleManifest,
-    extractDir: URL,
-    setupsDir: String,
-    courseID: UUID,
-    db: Database,
-    tally: inout ImportTally
-) async throws -> [String: String] {
-    var setupIDMap: [String: String] = [:]
-    for bundledSetup in manifest.testSetups {
-        let newSetupID = "setup_\(UUID().uuidString.lowercased().prefix(8))"
-        let newZipPath = setupsDir + "\(newSetupID).zip"
-
-        // Copy zip from bundle into testsetups dir.
-        let srcZip = extractDir.appendingPathComponent(bundledSetup.zipFilename)
-        try FileManager.default.copyItem(
-            at: srcZip,
-            to: URL(fileURLWithPath: newZipPath))
-
-        // Extract .ipynb if present (browser-mode setups).
-        var notebookPath: String?
-        if let nbData = extractNotebookFromZip(zipPath: newZipPath) {
-            let nbPath = setupsDir + "\(newSetupID).ipynb"
-            try nbData.write(to: URL(fileURLWithPath: nbPath))
-            notebookPath = nbPath
-        }
-
-        let setup = APITestSetup(
-            id: newSetupID,
-            manifest: bundledSetup.manifest,
-            zipPath: newZipPath,
-            notebookPath: notebookPath,
-            courseID: courseID
-        )
-        try await setup.save(on: db)
-        setupIDMap[bundledSetup.bundleID] = newSetupID
-        tally.testSetupsImported += 1
-    }
-    return setupIDMap
-}
-
-private func importBundledAssignments(
-    manifest: CourseBundleManifest,
-    setupIDMap: [String: String],
-    courseID: UUID,
-    db: Database,
-    tally: inout ImportTally
-) async throws {
-    for bundledAssign in manifest.assignments {
-        guard let setupID = setupIDMap[bundledAssign.testSetupBundleID] else { continue }
-        let newAssign = APIAssignment(
-            testSetupID: setupID,
-            title: bundledAssign.title,
-            slug: try await uniqueAssignmentSlug(title: bundledAssign.title, courseID: courseID, db: db),
-            dueAt: bundledAssign.dueAt,
-            startsAt: bundledAssign.startsAt,
-            visibility: bundledAssignmentVisibility(bundledAssign),
-            sortOrder: bundledAssign.sortOrder,
-            validationStatus: nil,  // not imported — requires re-validation
-            courseID: courseID
-        )
-        try await newAssign.save(on: db)
-        tally.assignmentsImported += 1
-    }
-}
-
-private func importBundledSubmissions(
-    manifest: CourseBundleManifest,
-    extractDir: URL,
-    subsDir: String,
-    idMaps: ImportIDMaps,
-    db: Database,
-    tally: inout ImportTally
-) async throws -> [String: String] {
-    let userIDMap = idMaps.userIDMap
-    let setupIDMap = idMaps.setupIDMap
-    var subIDMap: [String: String] = [:]
-    for bundledSub in manifest.submissions {
-        guard let setupID = setupIDMap[bundledSub.testSetupBundleID] else { continue }
-        let userID = userIDMap[bundledSub.userBundleID]
-
-        let srcFile = extractDir.appendingPathComponent(bundledSub.submissionFilename)
-        let ext = srcFile.pathExtension
-        let newSubID = "sub_\(UUID().uuidString.lowercased().prefix(8))"
-        let destName = ext.isEmpty ? "\(newSubID).bin" : "\(newSubID).\(ext)"
-        let newFilePath = subsDir + destName
-        try FileManager.default.copyItem(
-            at: srcFile,
-            to: URL(fileURLWithPath: newFilePath))
-
-        let sub = APISubmission(
-            id: newSubID,
-            testSetupID: setupID,
-            zipPath: newFilePath,
-            attemptNumber: bundledSub.attemptNumber,
-            status: SubmissionStatus.complete.rawValue,
-            filename: bundledSub.filename,
-            userID: userID,
-            kind: APISubmission.Kind.student
-        )
-        try await sub.save(on: db)
-        subIDMap[bundledSub.bundleID] = newSubID
-        tally.submissionsImported += 1
-    }
-    return subIDMap
-}
-
-private func importBundledResults(
-    manifest: CourseBundleManifest,
-    subIDMap: [String: String],
-    db: Database,
-    tally: inout ImportTally
-) async throws {
-    for bundledResult in manifest.results {
-        guard let subID = subIDMap[bundledResult.submissionBundleID] else { continue }
-        let newResultID = "res_\(UUID().uuidString.lowercased().prefix(8))"
-        let result = APIResult(
-            id: newResultID,
-            submissionID: subID,
-            collectionJSON: bundledResult.collectionJSON,
-            source: bundledResult.source
-        )
-        try await result.save(on: db)
-        tally.resultsImported += 1
-    }
 }
