+++ Sources/APIServer/Routes/Web/StudentCourseRoutes+History.swift
@@ -277,29 +277,17 @@ extension StudentCourseRoutes {
 
     @Sendable
     func studentAssignmentHistoryPage(req: Request) async throws -> View {
-        let viewer = try req.auth.require(APIUser.self)
-        guard viewer.isInstructor else {
-            throw WebAssignmentError.forbidden(action: "view student submission history")
-        }
-        let (course, student) = try await resolveCourseAndStudent(req: req)
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            assignment.courseID == course.id
-        else {
-            throw WebAssignmentError.notFound(resource: "Assignment '\(assignmentIDRaw)'")
-        }
+        let action = try await resolveStudentAssignmentAction(
+            req: req, action: "view student submission history")
+        let (course, student, assignment) = (action.course, action.student, action.assignment)
+        let assignmentIDRaw = assignment.publicID
 
-        let submissions: [APISubmission]
-        if let studentUUID = student.id {
-            submissions = try await APISubmission.query(on: req.db)
-                .filter(\.$testSetupID == assignment.testSetupID)
-                .filter(\.$userID == studentUUID)
-                .filter(\.$kind == APISubmission.Kind.student)
-                .sort(\.$submittedAt, .descending)
-                .all()
-        } else {
-            submissions = []
-        }
+        let submissions = try await APISubmission.query(on: req.db)
+            .filter(\.$testSetupID == assignment.testSetupID)
+            .filter(\.$userID == action.studentID)
+            .filter(\.$kind == APISubmission.Kind.student)
+            .sort(\.$submittedAt, .descending)
+            .all()
         let preferredResultBySubmissionID = try await preferredResultsBySubmissionID(
             for: submissions.compactMap(\.id),
             on: req.db
@@ -356,23 +344,14 @@ extension StudentCourseRoutes {
 
     @Sendable
     func retestStudentAssignment(req: Request) async throws -> Response {
-        let actor = try req.auth.require(APIUser.self)
-        guard actor.isInstructor else {
-            throw WebAssignmentError.forbidden(action: "retest student submissions")
-        }
-        let (course, student) = try await resolveCourseAndStudent(req: req)
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            assignment.courseID == course.id,
-            let studentID = student.id
-        else {
-            throw WebAssignmentError.notFound(resource: "Assignment '\(assignmentIDRaw)'")
-        }
+        let action = try await resolveStudentAssignmentAction(
+            req: req, action: "retest student submissions")
+        let (actor, student, assignment) = (action.actor, action.student, action.assignment)
+        let assignmentIDRaw = assignment.publicID
 
         let count = try await retestStudentSubmissionsForSetup(
             setupID: assignment.testSetupID,
-            studentUserID: studentID,
+            studentUserID: action.studentID,
             triggeredBy: actor.id,
             on: req.db,
             force: true
@@ -393,12 +372,7 @@ extension StudentCourseRoutes {
             on: req
         )
 
-        return req.redirect(
-            to: StudentCoursePaths.submissions(
-                courseCode: course.code,
-                urlToken: try student.requireURLToken()
-            )
-        )
+        return try redirectToStudentSubmissions(req: req, course: action.course, student: student)
     }
 
     // MARK: - POST /:courseCode/students/:urlToken/assignments/:assignmentID/reset-notebook
@@ -411,19 +385,10 @@ extension StudentCourseRoutes {
     /// redirect lands back here.
     @Sendable
     func resetStudentAssignmentNotebook(req: Request) async throws -> Response {
-        let actor = try req.auth.require(APIUser.self)
-        guard actor.isInstructor else {
-            throw WebAssignmentError.forbidden(action: "reset student notebooks")
-        }
-        let (course, student) = try await resolveCourseAndStudent(req: req)
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            assignment.courseID == course.id,
-            let studentID = student.id
-        else {
-            throw WebAssignmentError.notFound(resource: "Assignment '\(assignmentIDRaw)'")
-        }
+        let action = try await resolveStudentAssignmentAction(
+            req: req, action: "reset student notebooks")
+        let (actor, student, assignment) = (action.actor, action.student, action.assignment)
+        let assignmentIDRaw = assignment.publicID
         guard let setup = try await APITestSetup.find(assignment.testSetupID, on: req.db) else {
             throw WebAssignmentError.notFound(resource: "Test setup")
         }
@@ -441,7 +406,7 @@ extension StudentCourseRoutes {
         _ = try await ensureUserNotebookWorkingCopy(
             req: req,
             setupID: setup.id ?? assignment.testSetupID,
-            userID: studentID,
+            userID: action.studentID,
             fallbackSetup: setup,
             overwriteWith: starter
         )
@@ -450,12 +415,7 @@ extension StudentCourseRoutes {
             "student_notebook_reset assignment=\(assignmentIDRaw) student=\(student.username) by=\(actor.id?.uuidString ?? "nil")"
         )
 
-        return req.redirect(
-            to: StudentCoursePaths.submissions(
-                courseCode: course.code,
-                urlToken: try student.requireURLToken()
-            )
-        )
+        return try redirectToStudentSubmissions(req: req, course: action.course, student: student)
     }
 
     // MARK: - POST /:courseCode/students/:urlToken/assignments/:assignmentID/extension
@@ -467,18 +427,12 @@ extension StudentCourseRoutes {
             var note: String?
         }
 
-        let actor = try req.auth.require(APIUser.self)
-        guard actor.isInstructor else {
-            throw WebAssignmentError.forbidden(action: "grant deadline extensions")
-        }
-        let (course, student) = try await resolveCourseAndStudent(req: req)
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            assignment.courseID == course.id,
-            let assignmentUUID = assignment.id,
-            let studentUUID = student.id
-        else {
+        let action = try await resolveStudentAssignmentAction(
+            req: req, action: "grant deadline extensions")
+        let (actor, student) = (action.actor, action.student)
+        let assignmentIDRaw = action.assignment.publicID
+        let studentUUID = action.studentID
+        guard let assignmentUUID = action.assignment.id else {
             throw WebAssignmentError.notFound(resource: "Assignment '\(assignmentIDRaw)'")
         }
 
@@ -528,30 +482,19 @@ extension StudentCourseRoutes {
             on: req
         )
 
-        return req.redirect(
-            to: StudentCoursePaths.submissions(
-                courseCode: course.code,
-                urlToken: try student.requireURLToken()
-            )
-        )
+        return try redirectToStudentSubmissions(req: req, course: action.course, student: student)
     }
 
     // MARK: - POST /:courseCode/students/:urlToken/assignments/:assignmentID/extension/delete
 
     @Sendable
     func deleteStudentAssignmentExtension(req: Request) async throws -> Response {
-        let actor = try req.auth.require(APIUser.self)
-        guard actor.isInstructor else {
-            throw WebAssignmentError.forbidden(action: "revoke deadline extensions")
-        }
-        let (course, student) = try await resolveCourseAndStudent(req: req)
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            assignment.courseID == course.id,
-            let assignmentUUID = assignment.id,
-            let studentUUID = student.id
-        else {
+        let action = try await resolveStudentAssignmentAction(
+            req: req, action: "revoke deadline extensions")
+        let student = action.student
+        let assignmentIDRaw = action.assignment.publicID
+        let studentUUID = action.studentID
+        guard let assignmentUUID = action.assignment.id else {
             throw WebAssignmentError.notFound(resource: "Assignment '\(assignmentIDRaw)'")
         }
 
@@ -573,12 +516,7 @@ extension StudentCourseRoutes {
             )
         }
 
-        return req.redirect(
-            to: StudentCoursePaths.submissions(
-                courseCode: course.code,
-                urlToken: try student.requireURLToken()
-            )
-        )
+        return try redirectToStudentSubmissions(req: req, course: action.course, student: student)
     }
 
     // MARK: - POST /:courseCode/students/:urlToken/assignments/:assignmentID/grade-override
@@ -590,19 +528,11 @@ extension StudentCourseRoutes {
             var note: String?
         }
 
-        let actor = try req.auth.require(APIUser.self)
-        guard actor.isInstructor else {
-            throw WebAssignmentError.forbidden(action: "override grades")
-        }
-        let (course, student) = try await resolveCourseAndStudent(req: req)
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            assignment.courseID == course.id,
-            let studentUUID = student.id
-        else {
-            throw WebAssignmentError.notFound(resource: "Assignment '\(assignmentIDRaw)'")
-        }
+        let action = try await resolveStudentAssignmentAction(
+            req: req, action: "override grades")
+        let (actor, student, assignment) = (action.actor, action.student, action.assignment)
+        let assignmentIDRaw = assignment.publicID
+        let studentUUID = action.studentID
         let testSetupID = assignment.testSetupID
 
         let body = try req.content.decode(OverrideBody.self)
@@ -636,31 +566,18 @@ extension StudentCourseRoutes {
             on: req
         )
 
-        return req.redirect(
-            to: StudentCoursePaths.submissions(
-                courseCode: course.code,
-                urlToken: try student.requireURLToken()
-            )
-        )
+        return try redirectToStudentSubmissions(req: req, course: action.course, student: student)
     }
 
     // MARK: - POST /:courseCode/students/:urlToken/assignments/:assignmentID/grade-override/delete
 
     @Sendable
     func deleteStudentAssignmentGradeOverride(req: Request) async throws -> Response {
-        let actor = try req.auth.require(APIUser.self)
-        guard actor.isInstructor else {
-            throw WebAssignmentError.forbidden(action: "clear grade overrides")
-        }
-        let (course, student) = try await resolveCourseAndStudent(req: req)
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            assignment.courseID == course.id,
-            let studentUUID = student.id
-        else {
-            throw WebAssignmentError.notFound(resource: "Assignment '\(assignmentIDRaw)'")
-        }
+        let action = try await resolveStudentAssignmentAction(
+            req: req, action: "clear grade overrides")
+        let (student, assignment) = (action.student, action.assignment)
+        let assignmentIDRaw = assignment.publicID
+        let studentUUID = action.studentID
         let testSetupID = assignment.testSetupID
 
         if try await clearGradeOverride(
@@ -678,12 +595,7 @@ extension StudentCourseRoutes {
             )
         }
 
-        return req.redirect(
-            to: StudentCoursePaths.submissions(
-                courseCode: course.code,
-                urlToken: try student.requireURLToken()
-            )
-        )
+        return try redirectToStudentSubmissions(req: req, course: action.course, student: student)
     }
 }
 
@@ -726,6 +638,62 @@ extension StudentCourseRoutes {
             throw WebAssignmentError.notFound(resource: "Enrolled student")
         }
         return (course, student)
+    }
+
+    /// Everything the per-student assignment handlers resolve before doing
+    /// their real work: the authenticated instructor, the `(course, student)`
+    /// pair from `:courseCode` + `:urlToken`, and the `:assignmentID`
+    /// assignment verified to belong to that course.
+    fileprivate struct StudentAssignmentActionContext {
+        let actor: APIUser
+        let course: APICourse
+        let student: APIUser
+        let studentID: UUID
+        let assignment: APIAssignment
+    }
+
+    /// Shared resolve preamble for the seven per-student assignment handlers
+    /// (history page, retest, notebook reset, extension save/delete, grade
+    /// override save/delete).
+    ///
+    /// Note on the role check: these routes are registered under
+    /// `RoleMiddleware(required: .instructor)` (routes.swift), so the
+    /// `isInstructor` guard here is redundant — it is kept once, in this
+    /// helper, as defense-in-depth in case the route grouping ever changes.
+    /// `action` carries each handler's original forbidden-message wording.
+    ///
+    /// Error semantics match the guard chain each handler previously
+    /// inlined: `notFound("Assignment '<id>'")` when the assignment is
+    /// missing, belongs to a different course, or (unreachable for a
+    /// DB-loaded model) the student row has no id.
+    fileprivate func resolveStudentAssignmentAction(
+        req: Request, action: String
+    ) async throws -> StudentAssignmentActionContext {
+        let actor = try req.auth.require(APIUser.self)
+        guard actor.isInstructor else {
+            throw WebAssignmentError.forbidden(action: action)
+        }
+        let (course, student) = try await resolveCourseAndStudent(req: req)
+        let assignment = try await loadAssignment(req)
+        guard assignment.courseID == course.id, let studentID = student.id else {
+            throw WebAssignmentError.notFound(resource: "Assignment '\(assignment.publicID)'")
+        }
+        return StudentAssignmentActionContext(
+            actor: actor, course: course, student: student,
+            studentID: studentID, assignment: assignment)
+    }
+
+    /// Shared redirect epilogue: back to this student's per-course
+    /// submissions page.
+    fileprivate func redirectToStudentSubmissions(
+        req: Request, course: APICourse, student: APIUser
+    ) throws -> Response {
+        req.redirect(
+            to: StudentCoursePaths.submissions(
+                courseCode: course.code,
+                urlToken: try student.requireURLToken()
+            )
+        )
     }
 
     /// Bundles the per-table inputs that don't vary across rows.  Keeps
